requirements.txt

python-multipart
fastapi
uvicorn
motor
python-dotenv
websocket
Pillow
uuid
passlib
requests
bcrypt
python-jose
PyJwt